% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Bucket.owner}
\alias{Bucket.owner}
\title{Bucket.owner Object}
\usage{
Bucket.owner(entity = NULL, entityId = NULL)
}
\arguments{
\item{entity}{The entity, in the form group-groupId}

\item{entityId}{The ID for the entity}
}
\value{
Bucket.owner object
}
\description{
Bucket.owner Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The owner of the bucket. This is always the project team's owner group.
}
\seealso{
Other Bucket functions: \code{\link{Bucket.cors}},
  \code{\link{Bucket.lifecycle.rule.action}},
  \code{\link{Bucket.lifecycle.rule.condition}},
  \code{\link{Bucket.lifecycle.rule}},
  \code{\link{Bucket.lifecycle}},
  \code{\link{Bucket.logging}},
  \code{\link{Bucket.versioning}},
  \code{\link{Bucket.website}}, \code{\link{Bucket}},
  \code{\link{buckets.insert}},
  \code{\link{buckets.patch}}, \code{\link{buckets.update}}
}

