% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{BucketAccessControl}
\alias{BucketAccessControl}
\title{BucketAccessControl Object}
\usage{
BucketAccessControl(bucket = NULL, domain = NULL, email = NULL,
  entity = NULL, entityId = NULL, etag = NULL, id = NULL, role = NULL,
  selfLink = NULL)
}
\arguments{
\item{bucket}{The name of the bucket}

\item{domain}{The domain associated with the entity, if any}

\item{email}{The email address associated with the entity, if any}

\item{entity}{The entity holding the permission, in one of the following forms:}

\item{entityId}{The ID for the entity, if any}

\item{etag}{HTTP 1}

\item{id}{The ID of the access-control entry}

\item{role}{The access permission for the entity}

\item{selfLink}{The link to this access-control entry}
}
\value{
BucketAccessControl object
}
\description{
BucketAccessControl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An access-control entry.
}
\seealso{
Other BucketAccessControl functions: \code{\link{bucketAccessControls.insert}},
  \code{\link{bucketAccessControls.patch}},
  \code{\link{bucketAccessControls.update}}
}

