% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{BucketAccessControls}
\alias{BucketAccessControls}
\title{BucketAccessControls Object}
\usage{
BucketAccessControls(items = NULL)
}
\arguments{
\item{items}{The list of items}
}
\value{
BucketAccessControls object
}
\description{
BucketAccessControls Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An access-control list.
}

