% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{buckets.delete}
\alias{buckets.delete}
\title{Permanently deletes an empty bucket.}
\usage{
buckets.delete(bucket, ifMetagenerationMatch = NULL,
  ifMetagenerationNotMatch = NULL)
}
\arguments{
\item{bucket}{Name of a bucket}

\item{ifMetagenerationMatch}{Makes the return of the bucket metadata conditional on whether the bucket's current metageneration matches the given value}

\item{ifMetagenerationNotMatch}{Makes the return of the bucket metadata conditional on whether the bucket's current metageneration does not match the given value}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}
}
