% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{buckets.list}
\alias{buckets.list}
\title{Retrieves a list of buckets for a given project.}
\usage{
buckets.list(project, maxResults = NULL, pageToken = NULL,
  projection = NULL)
}
\arguments{
\item{project}{A valid API project identifier}

\item{maxResults}{Maximum number of buckets to return}

\item{pageToken}{A previously-returned page token representing part of the larger set of results to view}

\item{projection}{Set of properties to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_only, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}
}

