% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_objects.R
\name{Buckets}
\alias{Buckets}
\title{Buckets Object}
\usage{
Buckets(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{The list of items}

\item{nextPageToken}{The continuation token, used to page through large result sets}
}
\value{
Buckets object
}
\description{
Buckets Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A list of buckets.
}

