% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{objects.copy}
\alias{objects.copy}
\title{Copies an object to a destination in the same location. Optionally overrides metadata.}
\usage{
objects.copy(Object, sourceBucket, sourceObject, destinationBucket,
  destinationObject, ifGenerationMatch = NULL, ifGenerationNotMatch = NULL,
  ifMetagenerationMatch = NULL, ifMetagenerationNotMatch = NULL,
  ifSourceGenerationMatch = NULL, ifSourceGenerationNotMatch = NULL,
  ifSourceMetagenerationMatch = NULL, ifSourceMetagenerationNotMatch = NULL,
  projection = NULL, sourceGeneration = NULL)
}
\arguments{
\item{Object}{The \link{Object} object to pass to this method}

\item{sourceBucket}{Name of the bucket in which to find the source object}

\item{sourceObject}{Name of the source object}

\item{destinationBucket}{Name of the bucket in which to store the new object}

\item{destinationObject}{Name of the new object}

\item{ifGenerationMatch}{Makes the operation conditional on whether the destination object's current generation matches the given value}

\item{ifGenerationNotMatch}{Makes the operation conditional on whether the destination object's current generation does not match the given value}

\item{ifMetagenerationMatch}{Makes the operation conditional on whether the destination object's current metageneration matches the given value}

\item{ifMetagenerationNotMatch}{Makes the operation conditional on whether the destination object's current metageneration does not match the given value}

\item{ifSourceGenerationMatch}{Makes the operation conditional on whether the source object's generation matches the given value}

\item{ifSourceGenerationNotMatch}{Makes the operation conditional on whether the source object's generation does not match the given value}

\item{ifSourceMetagenerationMatch}{Makes the operation conditional on whether the source object's current metageneration matches the given value}

\item{ifSourceMetagenerationNotMatch}{Makes the operation conditional on whether the source object's current metageneration does not match the given value}

\item{projection}{Set of properties to return}

\item{sourceGeneration}{If present, selects a specific revision of the source object (as opposed to the latest version, the default)}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other Object functions: \code{\link{Object.metadata}},
  \code{\link{Object.owner}}, \code{\link{Object}},
  \code{\link{objects.insert}},
  \code{\link{objects.patch}}, \code{\link{objects.update}}
}

