% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{objects.insert}
\alias{objects.insert}
\title{Stores new data blobs and associated metadata.}
\usage{
objects.insert(Object, bucket, ifGenerationMatch = NULL,
  ifGenerationNotMatch = NULL, ifMetagenerationMatch = NULL,
  ifMetagenerationNotMatch = NULL, name = NULL, projection = NULL)
}
\arguments{
\item{Object}{The \link{Object} object to pass to this method}

\item{bucket}{Name of the bucket in which to store the new object}

\item{ifGenerationMatch}{Makes the operation conditional on whether the object's current generation matches the given value}

\item{ifGenerationNotMatch}{Makes the operation conditional on whether the object's current generation does not match the given value}

\item{ifMetagenerationMatch}{Makes the operation conditional on whether the object's current metageneration matches the given value}

\item{ifMetagenerationNotMatch}{Makes the operation conditional on whether the object's current metageneration does not match the given value}

\item{name}{Name of the object}

\item{projection}{Set of properties to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other Object functions: \code{\link{Object.metadata}},
  \code{\link{Object.owner}}, \code{\link{Object}},
  \code{\link{objects.copy}}, \code{\link{objects.patch}},
  \code{\link{objects.update}}
}

