% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\name{objects.update}
\alias{objects.update}
\title{Updates a data blob's associated metadata.}
\usage{
objects.update(Object, bucket, object, generation = NULL,
  ifGenerationMatch = NULL, ifGenerationNotMatch = NULL,
  ifMetagenerationMatch = NULL, ifMetagenerationNotMatch = NULL,
  projection = NULL)
}
\arguments{
\item{Object}{The \link{Object} object to pass to this method}

\item{bucket}{Name of the bucket in which the object resides}

\item{object}{Name of the object}

\item{generation}{If present, selects a specific revision of this object (as opposed to the latest version, the default)}

\item{ifGenerationMatch}{Makes the operation conditional on whether the object's current generation matches the given value}

\item{ifGenerationNotMatch}{Makes the operation conditional on whether the object's current generation does not match the given value}

\item{ifMetagenerationMatch}{Makes the operation conditional on whether the object's current metageneration matches the given value}

\item{ifMetagenerationNotMatch}{Makes the operation conditional on whether the object's current metageneration does not match the given value}

\item{projection}{Set of properties to return}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_write
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/devstorage.full_control, https://www.googleapis.com/auth/devstorage.read_write)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/storage/docs/json_api/}{Google Documentation}

Other Object functions: \code{\link{Object.metadata}},
  \code{\link{Object.owner}}, \code{\link{Object}},
  \code{\link{objects.copy}}, \code{\link{objects.insert}},
  \code{\link{objects.patch}}
}

