% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_functions.R
\docType{package}
\name{storage_googleAuthR}
\alias{storage_googleAuthR}
\alias{storage_googleAuthR-package}
\title{Cloud Storage JSON API
Lets you store and retrieve potentially-large, immutable data objects.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:48:40
filename: /Users/mark/dev/R/autoGoogleAPI/googlestoragev1beta2.auto/R/storage_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/devstorage.full_control
\item https://www.googleapis.com/auth/devstorage.read_only
\item https://www.googleapis.com/auth/devstorage.read_write
}
}

